Add luma_only mode to MPEG-1 decoder to reduce memory use
If you specify luma_only as true, the decoder will only allocate memory
for the Y buffer, not the Cr/Cb buffers and will avoid any code that
might try to modify cb.data or cr.data.  This is useful in an embedded
context where you're driving a monochrome screen or don't have enough
memory to decode color.